+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Don't persist jobs with the `JobScheduler`, if this device is weird and doesn't have the boot permission
 * List `support-v4` as dependency, because it's required, see #200
 * Make `Job.Params` public for better test support, see #201
+* Allow to suppress the `JobManagerCreateException`, see `JobApi.setForceAllowApi14(boolean)`
 
 ## 1.1.10 (2017-04-29)